% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{cls}
\alias{cls}
\title{cls}
\usage{
cls(hard = F)
}
\arguments{
\item{hard}{if TRUE, a full reset is performed. Usefull if something has messed up your terminal.}
}
\description{
clear screen (including scrollback buffer (as opposed to Ctrl-L))
}
