% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{colClasses}
\alias{colClasses}
\title{Expand template into \code{\link[base]{read.table}}'s \code{colClasses} argument}
\usage{
colClasses(fmt)
}
\arguments{
\item{fmt}{Format string to be expanded and then used as the
\code{colClasses} argument to \code{\link[base]{read.table}}.}
}
\value{
A vector that can be used as the \code{colClasses} argument
     to \code{\link[base]{read.table}}.
}
\description{
Expand template into \code{\link[base]{read.table}}'s \code{colClasses} argument
}
\details{
The format allows digits and the letters N, c, i, and n which will
expand into "NULL", "character", "integer", and "numeric".  The
digits say how many times the expansion of the character should be
included in the result.
}
\examples{
stopifnot(c("NULL", "NULL", "character", "character", "character",
     "integer", "numeric") == colClasses(c("2N3cin")))

}

