% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{count_points_per_interval}
\alias{count_points_per_interval}
\title{Count how many points fall into each interval}
\usage{
count_points_per_interval(pos, start, end)
}
\arguments{
\item{pos}{Points}

\item{start}{Left endpoints of interval}

\item{end}{Right endpoints of interval}
}
\value{
An integer vector of the same length as `start` and `end`
containing the number of points in `pos` that fall into each of the
intervals.
}
\description{
Count how many points fall into each interval
}
\examples{
count_points_per_interval(0:1, 2:1, 3:4)

}

