% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{double_quote}
\alias{double_quote}
\alias{single_quote}
\title{Surround string(s) with double/single quotes}
\usage{
double_quote(x)

single_quote(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Surrounds every element of `x' with double/single quotes,
escaping nested double quotes as needed.
}
\description{
Surround string(s) with double/single quotes
}

