% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.r
\name{dreamroc}
\alias{dreamroc}
\title{dreamroc}
\usage{
dreamroc(Truth, Prediction)
}
\arguments{
\item{Truth}{must be a named logical vector}

\item{Prediction}{must be a named numeric vector}
}
\description{
dreamroc
}
