% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{evalf}
\alias{evalf}
\title{Eval a piece of code supplied in a C-style string format}
\usage{
evalf(fmt, ...)
}
\arguments{
\item{fmt}{C-style string format in length-1 character vector}

\item{\dots}{values to be passed into \code{fmt}}
}
\value{
Returns the result of evaluating the code specified by
    \code{fmt} in the calling environment.
}
\description{
Eval a piece of code supplied in a C-style string format
}
\examples{
x <- 1; y <- 2
evalf("\%s + y", "x")

}

