% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{exit}
\alias{exit}
\title{Exit from running R job}
\usage{
exit(status = 0L, ...)
}
\arguments{
\item{status}{Exit status}

\item{\dots}{Other arguments that could also be passed to \code{\link[base]{quit}}}
}
\description{
Exit from running R job
}
\seealso{
\code{\link[base]{quit}}
}

