% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{find_first_match}
\alias{find_first_match}
\title{Find first matching candidate}
\usage{
find_first_match(candidates, x, ...)
}
\arguments{
\item{candidates}{Character vector of candidates to be matched
against `x'}

\item{x}{Character vector}

\item{\dots}{Further arguments passed on to \code{\link[base]{grepl}}}
}
\value{
#' Index of the element of `candidates' that matches the
element of `x' with the smallest index.  In case of ties, the
smallest of the tied indexes is returned.  If none of the
candidates match any of the elements in `x', NA is returned.
}
\description{
Find first matching candidate
}
\examples{
find_first_match(c("foo", "bar", "baz", "quux"), c("baz", "quux", "bar"))

}

