% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{find_matching_intervals}
\alias{find_matching_intervals}
\title{Find intervals containing a set of points}
\usage{
find_matching_intervals(pos, start, end)
}
\arguments{
\item{pos}{Double vector of positions}

\item{start}{Double vector with left endpoints of intervals}

\item{end}{Double vector with right endpoints of intervals}
}
\value{
A list with elements "position" and "interval".  Both are
    integer vectors containing indexes.  The "position" vector
    indexes the \code{pos} vector and the "interval" vector the
    \code{start} and \code{end} vectors.  The kth index in the
    "position" vector forms a pair with the kth index in the
    "interval" vector.  A pair of indexes (i,j) means that
    \code{pos[i]} belongs to the interval defined by
    \code{start[j]} and \code{end[j]}.
}
\description{
Find intervals containing a set of points
}
\examples{
x <- find_matching_intervals(c(2, 7), c(5, 0, 2), c(9, 1, 3))
stopifnot(identical(x$position, c(1L, 2L)))
stopifnot(identical(x$interval, c(3L, 1L)))
}

