% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{getmatches}
\alias{getmatches}
\title{getmatches}
\usage{
getmatches(pattern, txt)
}
\description{
get all regex-matches from a list of strings.
e.g.: unlist(getmatches("rs[0-9{3,16}", book)) will return all rs-ids from a
}
