% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{\%not_in\%}
\alias{\%not_in\%}
\title{Opposite of \code{\%in\%}}
\usage{
x \%not_in\% table
}
\arguments{
\item{x}{Values to be matched}

\item{table}{Values to be matched against}
}
\value{
Returns the logical negation of \code{\%in\%}.
}
\description{
Opposite of \code{\%in\%}
}
\seealso{
\code{\link[base]{\%in\%}}
}

