% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.r
\name{groupdiff}
\alias{groupdiff}
\title{groupdiff}
\usage{
groupdiff(dat, grp1, grp2, p = 0.001, wtest = "ttest")
}
\arguments{
\item{dat}{data matrix or similar}

\item{grp1}{and}

\item{grp2}{should be columns of `dat`.}

\item{wtest}{following tests are available:

"ttest" - log-2-t-test. `dat` should be signals

"wilcox" - Wilcoxon rank sum test. `dat` should be signals too

"fisher" - Exact fisher-test, `dat` should be T/F-table (specres)}
}
\value{
a list with a vector of peptides higher in group 1, a vector
of peptides higher in group 2 and, if fisher-test was done, a table
with the specific-counts (2x2-tables)
}
\description{
find differential peptides in 2 groupes.
}
