% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{gzhead}
\alias{gzhead}
\title{Output the first part of gzip-files}
\usage{
gzhead(files, n = 1L, simplify = TRUE)
}
\arguments{
\item{files}{Character vector with paths to gzip-files}

\item{n}{Number of lines to print}

\item{simplify}{If TRUE and `n = 1L', return result as character vector}
}
\value{
In case `simplify = FALSE', returns the first `n' lines of
each gzip-file in `files' as a list of character vectors.  If
`simplify = TRUE' and `n = 1' returns the results as a character
vector of the same length as `files'.
}
\description{
Output the first part of gzip-files
}

