% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{invert_map}
\alias{invert_map}
\title{Inverting a map}
\usage{
invert_map(map)
}
\arguments{
\item{map}{A named list}
}
\value{
A list representing the inverse of \code{map}
}
\description{
Given a named list, that is, a map, return the inverse map.
}
\details{
NA values are dropped during the inversion.
}
\examples{
map <- list(a = c("A", "C"), b = c("B", "C", NA))
invert_map(map)

}

