% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{is.directory}
\alias{is.directory}
\title{Test for directory}
\usage{
is.directory(fs)
}
\arguments{
\item{fs}{Filenames to be tested for directory status}
}
\value{
TRUE for directories, otherwise FALSE.
}
\description{
Test for directory
}

