% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{is_duplicated_in}
\alias{is_duplicated_in}
\title{Duplicated elements}
\usage{
is_duplicated_in(x)
}
\arguments{
\item{x}{Vector or list}
}
\value{
A logical vector of the same length as \code{x} that is
    TRUE if the corresponding element of \code{x} occurs more than
    once in \code{x}, and otherwise FALSE.  NAs always result in
    FALSE.
}
\description{
Duplicated elements
}
\seealso{
\code{\link[base]{duplicated}}
}

