% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{libs}
\alias{libs}
\title{libs}
\usage{
libs(pkg)
}
\arguments{
\item{pkg}{character vector containing package-names to load}
}
\description{
load multiple packages without startup messages but with colorfull progress
}
