% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{list2matrix}
\alias{list2matrix}
\title{list2matrix}
\usage{
list2matrix(lst, sym = NA)
}
\description{
create a matrix from a list of vectors (even with different lengths)
each list in one column but items will be alligned by content to create 
a feature-matrix. Sym can be NA (items will be used as they are)
or a string (such es "X" or "🗸")
}
