% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.r
\name{logttest}
\alias{logttest}
\title{logttest}
\usage{
logttest(dat, grp1, grp2, p = 0.001)
}
\description{
log2-t-Test for finding peptides, different in 2 groups.
`grp1` and `grp2` should be columns of `dat`.
Ex.: logttest(Signal[specificPeptide,], gCtr, gIll)
DEPRECATED. Use groupdiff instead
}
