% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{miscFun-package}
\alias{miscFun}
\alias{miscFun-package}
\title{Miscellaneous functions}
\description{
This is a collection of functions which, if not for this package, I
would have to rewrite or copy/paste for almost every R script that
I write.
}

