% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{order_intervals}
\alias{order_intervals}
\title{Order intervals in increasing order}
\usage{
order_intervals(start, end)
}
\arguments{
\item{start}{Left endpoints of interval}

\item{end}{Right endpoints of interval}
}
\value{
An integer vector `x' which, when used as an index to will
sort `start` and `end` into increasing order just as if
\code{\link[miscFun]{sort_intervals}} was used.
}
\description{
Order intervals in increasing order
}
\examples{
start <- 2:1
end <- 3:4
ord <- order_intervals(start, end)
start_sorted <- start[ord]
end_sorted <- end[ord]
old_order <- order(ord)
stopifnot(all(start_sorted[old_order] == start))
stopifnot(all(end_sorted[old_order] == end))

}
\seealso{
\code{\link[miscFun]{sort_intervals}}
}

