% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyprogress.r
\name{plapply}
\alias{plapply}
\title{plapply}
\usage{
plapply(
  X,
  FUN,
  ...,
  mc.cores = getOption("mc.cores", 4L),
  max.vector.size = getOption("max.vector.size", 1024L),
  mc.progress = "simple"
)
}
\description{
lapply & sapply für multicor mit automatischer (pretty)progress-bar.
}
