% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyprogress.r
\docType{class}
\name{ppbar-class}
\alias{ppbar-class}
\alias{ppbar}
\title{ppbar}
\arguments{
\item{maxval}{a vector of max values for progressbars. number of items = number of progressbars.}

\item{width}{width of prgressbar(s). Default is terminal-width as determined by tput}

\item{type}{one of 'txt', 'simple', 'small', 'pretty', 'plot', 'tk'}
}
\description{
Create one or many progressbar(s)
}

