% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{pr}
\alias{pr}
\alias{pr1}
\title{Print message}
\usage{
pr(...)

pr1(...)
}
\arguments{
\item{\dots}{Any number of arguments that could also be handed to
\code{\link[base]{cat}}}
}
\value{
\code{pr} prints its arguments pasted together followed by
a newline.

\code{pr1} prints its arguments pasted together but doesn't
add a trailing newline.
}
\description{
Print message
}
\examples{
pr("Finished iteration ", 3, " out of ", 10)

pr1("Progress: ")
for (i in 1:10) pr1(".")

}
\seealso{
\code{\link[base]{cat}}, \code{\link[base]{writeLines}}
}

