% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyprogress.r
\name{psapply}
\alias{psapply}
\title{psapply}
\usage{
psapply(
  X,
  FUN,
  ...,
  simplify = TRUE,
  USE.NAMES = TRUE,
  mc.cores = 1,
  max.vector.size = getOption("max.vector.size", 1024L),
  mc.progress = "small"
)
}
\description{
lapply & sapply für multicor mit automatischer (pretty)progress-bar.
}
