% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{read.fasta}
\alias{read.fasta}
\title{read.fasta}
\usage{
read.fasta(file, to.upper = FALSE)
}
\arguments{
\item{file}{file to read}

\item{to.upper}{convert everthing in input to upper case}
}
\description{
read FASTA-File
}
