% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{read}
\alias{read}
\title{read}
\usage{
read(input, sheet = 1, data.table = F, ...)
}
\description{
automagically read tables to data.frame or data.table.
Can read from textfiles (csv), even compressed (bz2, xz, gzip), excel-files.
Will read directly from http(s), ftp.
}
