% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{rev_map}
\alias{rev_map}
\title{Invert mapping}
\usage{
rev_map(map)
}
\arguments{
\item{map}{A named list representing a mapping}
}
\value{
A list representing the inverse of mapping of `map'
}
\description{
Given a mapping implemented as a named list, return the inverse
mapping, again implemented as a list.
}
\details{
NA values are dropped during the inversion.
}
\examples{
map <- list(a = c("A","C"), b = c("B","C",NA))
rev_map(map)

}

