% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.r
\name{roc}
\alias{roc}
\title{DEPRECATED
Create Points for a ROC-Curve. The result is an object with its own
print- and plot-functions. This is only provided for compatibility, use
dreamroc() instead.}
\usage{
roc(negatives, positives, noReduce = F)
}
\arguments{
\item{noReduce}{is no longer used}
}
\description{
DEPRECATED
Create Points for a ROC-Curve. The result is an object with its own
print- and plot-functions. This is only provided for compatibility, use
dreamroc() instead.
}
