% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{showxls}
\alias{showxls}
\title{showxls}
\usage{
showxls(mat, filename = NA)
}
\arguments{
\item{mat}{This may be a single object (anything that may be coerced to matrix). It will be exportet as a single-sheet-XLS.
Or it may be an environment. In this case, every object will be exportet as a single sheet in the resulting XLS.}

\item{filename}{Write to this file. If 'NA', it will be a tempfile() and directly opend in localc.}
}
\description{
export (and show) something as excel-file.
}
