% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic.r
\name{sunbeam}
\alias{sunbeam}
\title{SUNBEAM}
\usage{
sunbeam(specmat, grp_hi, grp_lo, verbose = F)
}
\description{
"simple unredundant thus necessarily binary-based essential attribute
minimization"
}
\details{
Algorithm for finding the best combination of specific
bound peptides, not bound in the controlgroup.
The aim is finding a combination of fewest possible peptides to tell
apart the sick from the healthy.
"Invented" during analysis of RA-ArrayIT (2009: JPT12[134]_..._1396 )
and SSc (2010: 01_Ssc)
For a given matrix of specificities ("specres") and two groups
(columns in specres) the best peptide-combination (rows in specres)
is returned. Has its own print- and plot-functions
}
