% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{write.fasta}
\alias{write.fasta}
\title{write.fasta}
\usage{
write.fasta(sq, file, wrap = 80, to.upper = FALSE)
}
\arguments{
\item{sq}{Sequences to be exported. They need names.}

\item{file}{Write to this file.}

\item{wrap}{Wrap sequence (not names) after this many chars}

\item{to.upper}{If TRUE, sequences will be converted to uppercase}
}
\description{
write FASTA-File. Beware! It is not sequential, so may be slow or bloaty for
larger Seqcence-vectors
}
