% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{ave2}
\alias{ave2}
\title{Apply function over level combinations of factors}
\usage{
ave2(x, factors, f, ...)
}
\arguments{
\item{x}{Vector which will be grouped according to `factors' and to
which `f' will be applied}

\item{factors}{List of factors, each of the same length as `x',
according to whose level combinations `x' will be grouped}

\item{f}{Function that will be applied to grouped subsets of `x'}

\item{\dots}{Further arguments to `f'}
}
\value{
A vector of the same length as `x'.  Every subset of `x'
     will get the value obtained by applying `f' to it.
}
\description{
Group a vector `x' into groups according to level combinations of
the factors in `factors' and apply `f' to every subset of `x'
separately.  The function \code{ave2} differs from
\code{\link[stats]{ave}} in the way arguments are specified and,
more importantly, in the fact that additional arguments to the
function `f' can be specified via `\dots'.
}
\examples{
fact1 <- rep(1:2, each = 4L)
fact2 <- rep(c("A","B"), length.out = 8L)
x <- 1:8
ave2(x, list(fact1, fact2), function(x, k) k * sum(x), 3)

}
\seealso{
\code{\link[stats]{ave}}
}

