% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{ave2}
\alias{ave2}
\title{Apply function over level combinations of factors}
\usage{
ave2(x, factors, f, ...)
}
\arguments{
\item{x}{Atomic vector}

\item{factors}{List of factors of the same length as \code{x}.}

\item{f}{Function}

\item{\dots}{Further arguments to \code{f}}
}
\value{
A vector of the same length as \code{x}.  Subsets of the
    result vector equal the result of applying \code{f} to the
    corresponding subsets of \code{x}.
}
\description{
Group a vector into subsets according to level
    combinations of one or several factors and apply a function
    separately to every subset.  This function differs from
    \code{\link[stats]{ave}} in three respects.  Firstly, factors
    have to be supplied in an explicit list.  Secondly, the
    function is a positional parameter.  Thirdly and most
    importantly, there is now a way to supply any number of
    additional arguments to the function via \code{\dots} as in the
    apply-family of functions.
}
\examples{
stopifnot(identical(
    ave2(c(1, 2, 3, 4), list(c(1, 1, 2, 2)),
        function(x, sep) {
            paste(x, rev(x), sep = sep)
        }, "-"),
    c("1-2", "2-1", "3-4", "4-3")))
}
\seealso{
\code{\link[stats]{ave}}
}

