% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{colClasses}
\alias{colClasses}
\title{Expand template into \code{\link[utils]{read.table}}s \code{colClasses} argument}
\usage{
colClasses(fmt)
}
\arguments{
\item{fmt}{Format string to be expanded and then used as the
\code{colClasses} argument to \code{\link[utils]{read.table}}.}
}
\value{
A vector that can be used as the \code{colClasses} argument
    to \code{\link[utils]{read.table}}.
}
\description{
Expand template into \code{\link[utils]{read.table}}s \code{colClasses} argument
}
\details{
The following format letters are available:
    \tabular{ll}{
      c\tab "character"\cr
      i\tab "integer"\cr
      l\tab "logical"\cr
      n\tab "numeric"\cr
      N\tab "NULL"\cr
      r\tab "raw"\cr
      x\tab "complex"\cr
    }
    The digits say how many times the expansion of the character
    should be included in the result.
}
\examples{
stopifnot(colClasses(c("N 2 c i n")) ==
    c("NULL", "character", "character", "integer", "numeric"))

}

