% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{find_duplicates}
\alias{find_duplicates}
\title{Extract duplicated rows from data frame.}
\usage{
find_duplicates(data, columns, select = NULL)
}
\arguments{
\item{data}{Data frame}

\item{columns}{Columns in \code{data} to be considered when looking
for duplicated rows.  Can be a character vector of column names
or a numeric vector giving the column indexes.}

\item{select}{Columns to include in resulting data frame.  By
default all columns will be included.}
}
\value{
A data frame containing all rows of \code{data} that had
    duplicates with respect to the columns specified in
    \code{columns}.
}
\description{
Extract all rows from a data frame that occur more than once when
compared in terms of the columns specified in \code{columns}.  Two
rows are duplicates with respect to the columns specified in
\code{columns} if they agree element-wise in all of the specified
columns.
}
\examples{
d <- read.table(text = "
x y z
1 2 3
1 3 4
1 2 4
", header = TRUE)

find_duplicates(d, "x")
find_duplicates(d, "x", c("y", "z"))
find_duplicates(d, c("x", "y"))
find_duplicates(d, c("x", "z"))

}
\seealso{
\code{\link{is_duplicated_in}}
}

