% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{find_duplicates}
\alias{find_duplicates}
\title{Extract duplicated rows from data frame.}
\usage{
find_duplicates(data, columns, select = NULL, sep = "\\t")
}
\arguments{
\item{data}{Data frame}

\item{columns}{Columns in \code{data} to be checked for duplicates.
Can be a character vector of column names or a numeric vector
giving the column indexes.}

\item{select}{Columns to include in resulting data frame}

\item{sep}{Separator used for row-wise pasting of values in
\code{cols} columns}
}
\value{
A data frame containing all rows of \code{data} that had
    duplicates with respect to the selected columns.
}
\description{
Extract all rows from a data frame that are duplicates with respect
to the specified columns.  Two rows are duplicates with respect to
the selected columns if they agree element-wise in all of the
selected columns.
}
\examples{
d <- read.table(text = "
x y z
1 2 3
1 3 4
1 2 4
", header = TRUE)

find_duplicates(d, "x")
find_duplicates(d, "x", c("y", "z"))
find_duplicates(d, c("x", "y"))
find_duplicates(d, c("x", "z"))

}
\seealso{
\code{\link[base]{duplicated}}
}

