% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{flatten_csv}
\alias{flatten_csv}
\title{Flatten list of strings of comma-separated elements}
\usage{
flatten_csv(x, sep = ",", fixed = TRUE)
}
\arguments{
\item{x}{Character vector of strings of separated elements}

\item{sep}{Separator used to separate elements}

\item{fixed}{If TRUE (default), separator is interpreted as a
constant string, otherwise it is considered a regular
expression in perl flavor}
}
\description{
Flatten list of strings of comma-separated elements
}
\details{
The separator can also be an arbitrary perl regular expression in
which case `fixed' should be set to \code{FALSE}.
}
\examples{
stopifnot(all(1:5 == flatten_csv(c("1,2", "3,4,5"))))
stopifnot(all(c("a", "b", "c", "d") == flatten_csv(c("a1b", "c23d"), sep = "\\\\\\\\d+", fixed = FALSE)))

}

