% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{make_pairs}
\alias{make_ordered_pairs}
\alias{make_pairs}
\alias{make_unordered_pairs}
\title{Create all [un]ordered pairs from two vectors}
\usage{
make_ordered_pairs(x, y = NULL)

make_unordered_pairs(x, y = NULL)
}
\arguments{
\item{x, y}{Atomic vectors to be paired off}
}
\value{
A data frame with column names "x" and "y" where every row
corresponds to a pair.
}
\description{
Create all [un]ordered pairs from two vectors
}
\examples{
make_ordered_pairs(1:3, 1:2)$x == c(1, 2, 3, 1, 2, 3)
make_ordered_pairs(1:3, 1:2)$y == c(1, 1, 1, 2, 2, 2)

make_unordered_pairs(1:3, 1:2)$x == c(1, 2, 3, 2, 3)
make_unordered_pairs(1:3, 1:2)$y == c(1, 1, 1, 2, 2)

}

