% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{match1of}
\alias{match1of}
\title{Find matching candidate}
\usage{
match1of(candidates, x, ...)
}
\arguments{
\item{candidates}{Character vector of candidates to be matched
against `x'}

\item{x}{Character vector}

\item{\dots}{Further arguments passed on to \code{\link[base]{grepl}}}
}
\value{
A character vector of the same length as `x' where element
i contains the first element of `candidates' that matches in
`x[i]', or NA if none of the candidates match.
}
\description{
Find matching candidate
}
\examples{
match1of(c("foo", "bar", "xxx", "baz", "quux"), c("foo", "bar", "quux"))

}

