% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{match_intervals}
\alias{match_intervals}
\title{Find intervals containing points}
\usage{
match_intervals(point, start, end)
}
\arguments{
\item{point}{Points}

\item{start}{Lower boundaries of intervals}

\item{end}{Upper boundaries of intervals}
}
\value{
A list of the same length as \code{point} where element
    \code{i} contains the indexes of closed intervals
    [\code{start}, \code{end}] containing \code{point[k]}.
    \code{NA} points result in an \code{NA} vector of indexes.
    Points that don't belong to any interval result in a length-0
    vector of indexes.
}
\description{
Find intervals containing points
}
\seealso{
\code{\link{find_matching_intervals}}
}

