% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{maybe}
\alias{maybe}
\title{Maybe apply function to value}
\usage{
maybe(f, x, y = x)
}
\arguments{
\item{f}{Function to be applied to `x'}

\item{x}{Object to which `f' should be applied}

\item{y}{Alternative value to return if applying `f' to `x' leads
to an error or a warning}
}
\value{
`f(x)' unless that results in an error or a warning,
otherwise `y'.
}
\description{
Maybe apply function to value
}
\details{
Errors or warnings caused by evaluating `x' will not be caught.
The purpose of \code{maybe} is only to catch errors and warning
caused by evaluating `f(x)'.
}
\examples{
x <- "12"
y <- "foo"
maybe(as.integer, x)
maybe(as.integer, y)
maybe(as.integer, y, 0L)
tryCatch(maybe(as.integer, stop(), 0L),
         error = function(e) pr("Evaluating `x' caused an error."))

}

