\name{miscFun-package}
\alias{miscFun-package}
\alias{miscFun}
\docType{package}
\title{
  Miscellaneous functions
}
\description{
  This is a collection of functions which, if not for this package, I
  would have to rewrite or copy/paste for almost every R script that I
  write.
}
\details{
  \tabular{ll}{
    Package: \tab miscFun\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2014-12-19\cr
    License: \tab GPL (>= 2)\cr
}
library(miscFun)
}
\author{
  Clemens Baumbach

  Maintainer: Clemens Baumbach <clemens.baumbach@helmholtz-muenchen.de>
}
