% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{nfields}
\alias{nfields}
\title{Return number of columns in file}
\usage{
nfields(con, sep = "\\t")
}
\arguments{
\item{con}{File or connection for which columns should be counted}

\item{sep}{Separator used to separate columns in \code{file}}
}
\description{
Return number of columns in file
}
\examples{
nfields("data.csv", sep = ",")

}

