% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{nsubexp}
\alias{nsubexp}
\title{Count number of parenthesized subexpressions}
\usage{
nsubexp(x)
}
\arguments{
\item{x}{Character vector}
}
\description{
Count number of parenthesized subexpressions
}
\examples{
nsubexp("a_[(string)]_with_(2)_parenthesized_(sub)expressions")

}

