% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_find_matching_intervals}
\alias{rcpp_find_matching_intervals}
\title{Find intervals containing a set of points}
\usage{
rcpp_find_matching_intervals(pos, start, end)
}
\arguments{
\item{pos}{Double vector of positions}

\item{start}{Double vector with left endpoints of intervals}

\item{end}{Double vector with right endpoints of intervals}
}
\value{
A list with elements "position" and "interval".  Both are integer
vectors containing indexes.  The "position" vector indexes the
\code{pos} vector and the "interval" vector the \code{start} and
\code{end} vectors.  The kth index in the "position" vector forms
a pair with the kth index in the "interval" vector.  A pair of
indexes (i,j) means that \code{pos[i]} belongs to the interval
defined by \code{start[j]} and \code{end[j]}.
}
\description{
Find intervals containing a set of points
}
\examples{
rcpp_find_matching_intervals(c(1.0, 2.0, 3.0), c(2.0, 3.0), c(4.0, 5.0))

}

