% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{same_length}
\alias{same_length}
\title{Test whether all arguments have the same length}
\usage{
same_length(...)
}
\arguments{
\item{\dots}{Objects that will be compared in terms of their length}
}
\value{
TRUE if all objects have the same length, otherwise FALSE.
}
\description{
Test whether all arguments have the same length
}
\examples{
stopifnot(
    same_length(NA, 0),
    same_length(NULL, integer()),
    same_length(list(), integer()),
    same_length(letters, letters, letters),
    !same_length(letters[1], letters[1:2], letters[1:3]))

}

