% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\name{showxls}
\alias{showxls}
\title{showxls}
\usage{
showxls(mat, filename = NA)
}
\arguments{
\item{mat}{This may be:
- a single object (anything that may be coerced to a data frame). It will be exportet as a single-sheet-XLSX.
- an environment or list. Every object in the list which can be coerced to a data.frame will be exportet as a single sheet in the resulting XLSX.}

\item{filename}{Write to this file. If 'NA', it will be a tempfile() and directly opend in localc.}
}
\description{
export (and show) something as excel-file.
}
