% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFun.R
\name{sort_intervals}
\alias{sort_intervals}
\title{Sort intervals in increasing order}
\usage{
sort_intervals(start, end)
}
\arguments{
\item{start}{Left endpoints of interval}

\item{end}{Right endpoints of interval}
}
\value{
A list with two vectors named "start" and "end" of the same
length as `start` and `end` and representing the left and right
endpoints of the sorted intervals.
}
\description{
Sort intervals in increasing order
}
\examples{
sort_intervals(2:1, 3:4)

}

